Make device-orientation readable
Hello micku7zu,
You have done wonderful work by creating vanilla-tilt.js. It is used by millions of developers like me. I just read your code and found that deviceorientation is not readable because it is not in the camel case and also not used any hyphen or underscore to identify it. It is very respectful project, that's why I created a pull request for it. Please consider it. Thank you.